Syntactical Sugar for Functional Accessors
{"pullable":true,"stable":true,"issues":[{"feature":true,"title":"Functional Syntactic Sugar","desc":"When 
there are no ambiguities, trying to get (either as a class variable or via a corresponding Get method) an 
object's property which is also the name of one of its classes functions, it will return a new ServerEvent 
instance having that object as the Source and using that method name. Any parameters passed to the Get method 
will be used as the parameters for the ServerEvent (or none will be passed in if the class variable style of 
getting a property is used instead of the Get method.) An article clarifying this with examples as well as 
further discussion is forthcoming. ","secret":"","order":"","internal":false}]}