Listener
{"pullable":true,"stable":true,"issues":[{"feature":true,"title":"Listener","desc":"The 
Listener Control allows a developer to listen to some data source and 
when a change in it is detected, the object's Update Event is 
automatically called for you. Thus, it is NOLOH's implementation of 
server push or \"Comet\" technology. Please read the upcoming article 
for more information.","secret":"","order":"","internal":false}]}